Refactor RAG service into modular components
