+++ rag_demo/app/app/services/rag/__init__.py
@@ -1,0 +1,4 @@
+from .service import RagService
+
+__all__ = ["RagService"]
+
